requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
SQLAlchemy==2.0.38
Werkzeug==2.2.2
WTForms==3.2.1
Jinja2==3.1.6
Flask-Mail==0.9.1
python-dotenv==1.0.1
Flask-Migrate==4.0.7
gunicorn==21.2.0
transformers==4.30.2
torch==2.0.1